Pull out the pprof server
