+++ common/pprof/server.go
@@ -1,4 +1,4 @@
-package apiserver
+package pprof
 
 import (
 	"fmt"
