+++ disperser/apiserver/server.go
@@ -17,6 +17,7 @@ import (
 	pb "github.com/Layr-Labs/eigenda/api/grpc/disperser"
 	"github.com/Layr-Labs/eigenda/common"
 	"github.com/Layr-Labs/eigenda/common/healthcheck"
+	"github.com/Layr-Labs/eigenda/common/pprof"
 	"github.com/Layr-Labs/eigenda/core"
 	"github.com/Layr-Labs/eigenda/core/auth"
 	"github.com/Layr-Labs/eigenda/core/meterer"
@@ -819,7 +820,7 @@ func (s *DispersalServer) GetRateConfig() *RateConfig {
 }
 
 func (s *DispersalServer) Start(ctx context.Context) error {
-	pprofProfiler := NewPprofProfiler(s.serverConfig.PprofHttpPort, s.logger)
+	pprofProfiler := pprof.NewPprofProfiler(s.serverConfig.PprofHttpPort, s.logger)
 	if s.serverConfig.EnablePprof {
 		go pprofProfiler.Start()
 		s.logger.Info("Enabled pprof for disperser apiserver", "port", s.serverConfig.PprofHttpPort)
